learnt about export and import and created separate files and folders for components
