+++ index.html
@@ -11,6 +11,6 @@
         <h1>Not Rendered</h1>
     </div>
 
-    <script type="module" src="./app.js"></script>
+    <script type="module" src="./src/app.js"></script>
 </body>
 </html>
