+++ src/app.js
@@ -1,0 +1,23 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import Header from "./components/Header";
+import Body from "./components/Body";
+
+/**
+ * Components of our Food Ordering App
+ * Header -> logo, nav items(home,about etc)
+ * Body -> search, rest containers->rest cards->img,name,cuisines,star rating,delivery time
+ * Footer -> copywright, social media links, information, address
+ */
+
+const AppLayout = () => {
+    return(
+        <div className="app">
+            <Header/>
+            <Body/>
+        </div>
+    )
+}
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(<AppLayout/>);
